# TrelloClone-frontend/src/api/models/Board.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Column } from './Column';
export type Board = {
    id?: string;
    name?: string | null;
    columns?: Array<Column> | null;
};