# TrelloClone-frontend/src/api/models/BoardDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ColumnDto } from './ColumnDto';
export type BoardDto = {
    id?: string;
    name?: string | null;
    columns?: Array<ColumnDto> | null;
};