# TrelloClone-frontend/src/api/models/ColumnDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TaskDto } from './TaskDto';
export type ColumnDto = {
    id?: string;
    name?: string | null;
    tasks?: Array<TaskDto> | null;
};