# TrelloClone-frontend/src/api/models/TaskDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TaskType } from './TaskType';
export type TaskDto = {
    id?: string;
    title?: string | null;
    description?: string | null;
    type?: TaskType;
    columnId?: string;
};