# TrelloClone.Api/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "TrelloClone.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/Board": {
      "get": {
        "tags": [
          "Board"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Board"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Board"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Board"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Board"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          }
        }
      }
    },
    "/api/Board/{id}": {
      "get": {
        "tags": [
          "Board"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Columns": {
      "get": {
        "tags": [
          "Columns"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ColumnDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ColumnDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ColumnDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Columns"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Column"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Column"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Column"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Column"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Column"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Column"
                }
              }
            }
          }
        }
      }
    },
    "/api/Columns/{id}": {
      "get": {
        "tags": [
          "Columns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ColumnDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColumnDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColumnDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TaskItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/Tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "TrelloClone.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Board": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Column"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BoardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Column": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "boardId": {
            "type": "string",
            "format": "uuid"
          },
          "board": {
            "$ref": "#/components/schemas/Board"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ColumnDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "columnId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "TaskItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "taskType": {
            "$ref": "#/components/schemas/TaskType"
          },
          "columnId": {
            "type": "string",
            "format": "uuid"
          },
          "column": {
            "$ref": "#/components/schemas/Column"
          }
        },
        "additionalProperties": false
      },
      "TaskType": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      }
    }
  },
  "tags": [
    {
      "name": "Board"
    },
    {
      "name": "Columns"
    },
    {
      "name": "Tasks"
    },
    {
      "name": "TrelloClone.Api"
    }
  ]
}